gnuboy 尚未跑通

+++ main/debug.h
@@ -2,6 +2,7 @@
 #define ONCE_H_DEBUG
 
 #include <stdio.h>
+#include "psram.h"
 
 #define pf(...) printf(__VA_ARGS__) ;printf("\n") ;
 #define dd printf("@%d\n", __LINE__) ;
@@ -9,12 +10,12 @@
 #define dm(msg) printf("%s: %dKB\n", msg, esp_get_free_heap_size()/1024);
 #define dp(p)   printf(#p"@%p\n", p) ;
 #define ds(s)   printf(#s"=%s @%d\n", s, __LINE__) ;
-#define dn(v)   printf(#v"=%d @%d\n", v, __LINE__) ;
+#define dn(v)   printf(#v"=%d @%d\n", (int)v, __LINE__) ;
 #define dn64(v)   printf(#v"=%lld\n", v) ;
-#define dn2(v1,v2)              printf(#v1"=%d, "#v2"=%d @%d\n", v1, v2, __LINE__) ;
-#define dn3(v1,v2,v3)           printf(#v1"=%d, "#v2"=%d, "#v3"=%d @%d\n", v1, v2, v3, __LINE__) ;
-#define dn4(v1,v2,v3,v4)        printf(#v1"=%d, "#v2"=%d, "#v3"=%d, "#v4"=%d @%d\n", v1, v2, v3, v4, __LINE__) ;
-#define dn5(v1,v2,v3,v4,v5)     printf(#v1"=%d, "#v2"=%d, "#v3"=%d, "#v4"=%d, "#v5"=%d @%d\n", v1, v2, v3, v4, v5, __LINE__) ;
+#define dn2(v1,v2)              printf(#v1"=%d, "#v2"=%d @%d\n", (int)v1, (int)v2, __LINE__) ;
+#define dn3(v1,v2,v3)           printf(#v1"=%d, "#v2"=%d, "#v3"=%d @%d\n", (int)v1, (int)v2, (int)v3, __LINE__) ;
+#define dn4(v1,v2,v3,v4)        printf(#v1"=%d, "#v2"=%d, "#v3"=%d, "#v4"=%d @%d\n", (int)v1, (int)v2, (int)v3, (int)v4, __LINE__) ;
+#define dn5(v1,v2,v3,v4,v5)     printf(#v1"=%d, "#v2"=%d, "#v3"=%d, "#v4"=%d, "#v5"=%d @%d\n", (int)v1, (int)v2, (int)v3, (int)v4, (int)v5, __LINE__) ;
 
 #define dfunc   printf("%s()@%d\n", __FUNCTION__, __LINE__) ;
